+++ lib/src/services/index.dart
@@ -9,3 +9,4 @@ export 'concretes/custom/custom_version_compare_manager.dart';
 export 'concretes/version_dialog_manager.dart';
 export 'concretes/android/android_version_convert_manager.dart';
 export 'concretes/ios/ios_version_convert_manager.dart';
+export 'concretes/http_remote_data_manager.dart';