+++ lib/src/utils/enums/app_platform.dart
@@ -1,1 +1,12 @@
-enum AppPlatform { android, ios, huawei, invalid }
+/// Enum representing different application platforms.
+///
+/// The available platforms are:
+/// - `android`: Represents the Android platform.
+/// - `ios`: Represents the iOS platform.
+/// - `huawei`: Represents the Huawei platform.
+/// - `web`: Represents web applications.
+/// - `windows`: Represents the Windows platform.
+/// - `linux`: Represents the Linux platform.
+/// - `macOs`: Represents the macOS platform.
+/// - `invalid`: Represents an invalid or unsupported platform.
+enum AppPlatform { android, ios, huawei, web, windows, linux, macOs, invalid }