+++ lib/src/utils/index.dart
@@ -5,6 +5,6 @@ export 'enums/check_version_loading_type.dart';
 export 'helpers/regexp_helper.dart';
 export 'messages/comparator_error_message.dart';
 export 'messages/comparator_info_message.dart';
-export 'mixins/platform_decider_mixin.dart';
+export 'mixins/platform/platform_decider_mixin.dart';
 export 'mixins/version_compare_mixin.dart';
 export 'mixins/version_dialog_mixin.dart';